* Equalizer3Bands: update AudioInfo

* VolumeStream: correct reported AudioInfo

* Equalizer3Bands: notifications

* ModifyingStream: add setOutput and setStream with notification

* PWM config carification

* DRAFT MP3 Shine Encoder

* shine (performance) test case

* USBAudio: new Buffer concept, standardize bool resize(size_t)

* USB: Using NBuffer

* USB correction for ESP32

* Compile error ESP32 V 2.017

* USB: RP2040 Support

* USBDescriptor: getInterfaceCount()

* Delete Sandbox USB

* usb-rx-quality.ino example

* New QualityAnalysisStream

* QualityAnalysisStream

* codecs: fix Ogg Opus container generation (#2368)

* codecs: fix Ogg Opus container generation

* codecs: use 64-bit Ogg granule positions

* fix: harden RTSP request parsing against buffer overflows (#2352)

* fix: V-001 security vulnerability

Automated security fix generated by OrbisAI Security

* fix: add buffer-length check in RTSPSession.h

The RTSP session handler copies incoming request data into the mCurRequest buffer using memcpy without verifying that aRequestSize does not exceed the buffer's allocated capacity

* fix: address code review comments on RTSP buffer hardening

- Remove overlapping-buffer snprintf in parseClientPorts: the second
  snprintf(CP, …, "%s", eq) had CP==dest and eq pointing into CP as
  src, which is undefined behaviour. Replace with in-place strstr/atoi
  on eq directly (no copy needed).
- Drop tests/test_invariant_RTSPSession.h: the file exercised a
  self-contained phantom implementation (safe_rtsp_copy /
  RTSPRequestBuffer with RTSP_BUFFER_CAPACITY=4096) unconnected to the
  real RTSPSession.h code (RTSP_BUFFER_SIZE=10000). Shipping it would
  provide false confidence.
- Update PR #2352 description to reflect hardening rather than
  CRITICAL/RCE framing.


---------


---------

+++ src/AudioTools/Concurrency/Zephyr/QueueZephyr.h
@@ -58,7 +58,7 @@ class QueueZephyr {
   void setWriteMaxWait(TickType_t ticks) { write_max_wait = ticks; }
 
   /// (Re-)defines the size
-  bool resize(int size) {
+  bool resize(size_t size) {
     bool result = true;
     if (size != queue_size) {
       end();
